refactor: move the NewSubscriberModal component to the subscribers directory
